+++ main.go
@@ -62,7 +62,7 @@ func main() {
 
 	app := cli.NewApp()
 	app.Name = "Pumba"
-	app.Version = "0.1.5"
+	app.Version = "0.1.6"
 	app.Usage = "Pumba is a resiliency tool that helps applications tolerate random Docker container failures."
 	app.Before = before
 	app.Commands = []cli.Command{
@@ -156,9 +156,9 @@ func before(c *cli.Context) error {
 		log.AddHook(&slackrus.SlackrusHook{
 			HookURL:        c.GlobalString("slackhook"),
 			AcceptedLevels: slackrus.LevelThreshold(log.GetLevel()),
-			Channel:        "#slack-testing",
+			Channel:        c.GlobalString("slackchannel"),
 			IconEmoji:      ":boar:",
-			Username:       c.GlobalString("slackchannel"),
+			Username:       "pumba_bot",
 		})
 	}