+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.1.7]
 ### Added
 ### Fixed
+- Report Pumba events to Slack
 
 ## [0.1.6] - 2016-04-25
 ### Added
@@ -12,7 +13,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Use Docker Label `com.gaiaadm.pumba.skip` to make Pumba ignoring the container. Avoid abusing it though.
 ### Fixed
 - Pumba runs as `pumba:pumba` user, instead of `root`
-- Report Pumba events to Slack
 
 ## [0.1.5] - 2016-04-13
 ### Added